+++ main.go
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	"github.com/golang/glog"
-	my_proxy "github.com/jthomperoo/simple-proxy/proxy"
+	"github.com/jthomperoo/simple-proxy/proxy"
 )
 
 var (
@@ -37,7 +37,9 @@ func main() {
 	var port string
 	flag.StringVar(&port, "port", "8888", "proxy port to listen on")
 	var socks5 string
-	flag.StringVar(&socks5, "socks5", "", "SOCKS5 proxy for tunneling")
+	flag.StringVar(&socks5, "socks5", "", "SOCKS5 proxy for tunneling, not used if not provided")
+	var socks5Auth string
+	flag.StringVar(&socks5Auth, "socks5-auth", "", "basic auth for socks5, format 'username:password', no auth if not provided")
 	var certPath string
 	flag.StringVar(&certPath, "cert", "", "path to cert file")
 	var keyPath string
@@ -65,26 +67,42 @@ func main() {
 		glog.Fatalf("If using HTTPS protocol --cert and --key are required\n")
 	}
 
-	my_proxy.Socks5 = socks5
+	var socks5Forward *proxy.Socks5Forward
+	if socks5 != "" {
+		socks5Forward = &proxy.Socks5Forward{
+			Address: socks5,
+		}
+		if socks5Auth != "" {
+			parts := strings.Split(socks5Auth, ":")
+			if len(parts) < 2 {
+				glog.Fatalf("Invalid socks5 basic auth provided, must be in format 'username:password', auth: %s\n", basicAuth)
+			}
+
+			socks5Forward.Username = &parts[0]
+			socks5Forward.Password = &parts[1]
+		}
+	}
 
 	var handler http.Handler
 	if basicAuth == "" {
-		handler = &my_proxy.ProxyHandler{
-			Timeout:    time.Duration(timeoutSecs) * time.Second,
-			LogAuth:    logAuth,
-			LogHeaders: logHeaders,
+		handler = &proxy.ProxyHandler{
+			Timeout:       time.Duration(timeoutSecs) * time.Second,
+			LogAuth:       logAuth,
+			LogHeaders:    logHeaders,
+			Socks5Forward: socks5Forward,
 		}
 	} else {
 		parts := strings.Split(basicAuth, ":")
 		if len(parts) < 2 {
 			glog.Fatalf("Invalid basic auth provided, must be in format 'username:password', auth: %s\n", basicAuth)
 		}
-		handler = &my_proxy.ProxyHandler{
-			Timeout:    time.Duration(timeoutSecs) * time.Second,
-			Username:   &parts[0],
-			Password:   &parts[1],
-			LogAuth:    logAuth,
-			LogHeaders: logHeaders,
+		handler = &proxy.ProxyHandler{
+			Timeout:       time.Duration(timeoutSecs) * time.Second,
+			Username:      &parts[0],
+			Password:      &parts[1],
+			LogAuth:       logAuth,
+			LogHeaders:    logHeaders,
+			Socks5Forward: socks5Forward,
 		}
 	}